refactor(functions): split triggers into separate files
